# user/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "User service API",
    "version": "1.0.0",
    "title": "Melting Pot - User"
  },
  "host": "localhost:5000",
  "tags": [
    {
      "name": "User",
      "description": "Everything about user"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Add a new user",
        "description": "",
        "operationId": "createUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be created",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created",
            "schema": {
              "$ref": "#/definitions/CreateUserResponse"
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "summary": "List users",
        "description": "List all users",
        "operationId": "listUsers",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of users successfully retrieved",
            "schema": {
              "$ref": "#/definitions/ListUsers"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CreateUserRequest": {
      "type": "object",
      "required": [
        "fullname"
      ],
      "properties": {
        "fullname": {
          "type": "string",
          "example": "John Doe"
        }
      }
    },
    "CreateUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "some-id"
        }
      }
    },
    "ListUsers": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/User"
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "fullname": {
          "type": "string",
          "example": "John Doe"
        },
        "cardStatus": {
          "type": "string",
          "enum": [
            "active",
            "closed",
            "frozen"
          ]
        },
        "walletBalance": {
          "type": "number",
          "example": 500
        },
        "createdAt": {
          "type": "string",
          "example": "2020-11-04T15:43:22.085Z"
        },
        "updateAt": {
          "type": "string",
          "example": "2020-11-04T15:43:22.085Z"
        }
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "example": 500
        },
        "type": {
          "type": "string",
          "example": "Internal"
        },
        "message": {
          "type": "string",
          "example": "Internal server error"
        }
      }
    }
  }
}